refactor: 디렉토리 내 테스트 코드와 소스 코드가 함께 위치하도록 변경
