+++ docs/MISSION_GOAL.md
@@ -6,8 +6,7 @@
   - [ ] 로또 게임에 대한 한 줄 소개를 추가한다.
   - [ ] 앱 작동 과정에 대한 gif를 첨부한다.
   - [ ] 게임에 대한 기능 목록을 추가한다.
-  - [ ] 게임에 대한 플로우 차트를 추가한다.
-- [ ] 기능 목록을 업데이트 한다.
+  - [ ] 프로젝트에 대한 아키텍처를 ARCHITECTURE.md에 추가한다.
 - [ ] 테스트를 작성하는 이유에 대해 본인의 경험을 토대로 정리한다.
 
 ## 객체 분리 연습
