+++ src/domain/confirmWinningInfo/rateOfReturnCalculation/rateOfReturnCalculation.module.js
