+++ src/validations/bonusNumberValidation/bonusNumberValidation.module.js
@@ -1,8 +1,8 @@
-import { LOTTO_RULES } from '../constants/lottoGame.js';
-import AppError from '../error/customErrors/AppError.js';
+import { LOTTO_RULES } from '../../constants/lottoGame.js';
+import AppError from '../../error/AppError/AppError.module.js';
 
 /**
- * @type {import('../utils/jsDoc.js').BonusNumberValidationType}
+ * @type {import('../../utils/jsDoc.js').BonusNumberValidationType}
  */
 export const BONUS_NUMBER_VALIDATION_TYPES = {
   bonusNumberRange: {
@@ -21,7 +21,7 @@ export const BONUS_NUMBER_VALIDATION_TYPES = {
 };
 
 /**
- * @param {import('../utils/jsDoc.js').WinningLottoInfo} winningLottoInfo - 보너스 번호와 당첨 번호가 있는 객체
+ * @param {import('../../utils/jsDoc.js').WinningLottoInfo} winningLottoInfo - 보너스 번호와 당첨 번호가 있는 객체
  * @throws {AppError} 유효성을 만족하지 않을 경우 에러 발생
  * @returns {void}
  */
